Make 'clean' builds work
When one tries to build miriad when it has been freshly unpacked into a         
clean directory, there is no $(MIRARCH)/aux/prog.inclist file.                  
This contains the explicit dependencies needed to trigger rebuilds              
if header files change, etc - e.g.                                              
 $ grep uvlist prog.inclist                                                     
 $(MIRBIND)/uvlist : $(MIRINCD)/linux64/maxdim.h                                
 $(MIRBIND)/uvlist : $(MIRINCD)/mirconst.h                                      
 $(MIRBIND)/uvlist : $(MIRPROGD)/uvlist.h                                       

But these dependencies have another function -                                  
they provide an explicit target like '$(MIRBIND)/uvlist' that gives             
the pattern rules something to latch onto. If the file is empty,                
then the pattern rules are unable to trigger.                                   

This list is built from files in aux/prog.inc/, so this patch makes             
that dependency explicit. Make checks if the directory is up to date            
and somehow, not quite sure how, this triggers the pattern rules                
that run ratty, do the dependencies and finally build the programs.             

This file is included by prog/GNUmakefile _but_ it uses -include,               
which will make no noise if the file to include doesn't exist.                  
It seems that up until now, either the build system has been relying on         
make rebuilding that file (make checks an included file to see if it needs      
to be remade), or maybe it never worked for a clean build.                      
Add a note about the behaviour of -include(), for future readers.